Break arithmetic.cpp up into separate cpp files (one for each callable). This should help parallel builds run more smoothly and avoid running out of memory during compilation on the CI systems and anywhere else where that's a major concern.
