+++ include/dynd/kernels/compound_add_kernel.hpp
@@ -6,6 +6,7 @@
 #pragma once
 
 #include <dynd/kernels/base_kernel.hpp>
+#include <dynd/kernels/base_strided_kernel.hpp>
 
 namespace dynd {
 namespace nd {
@@ -19,14 +20,12 @@ namespace nd {
     typedef DstType dst_type;
     typedef Src0Type src0_type;
 
-    void single(char *dst, char *const *src)
-    {
+    void single(char *dst, char *const *src) {
       *reinterpret_cast<dst_type *>(dst) += *reinterpret_cast<src0_type *>(src[0]);
     }
 
     void strided(char *dst, std::intptr_t dst_stride, char *const *src, const std::intptr_t *src_stride,
-                 std::size_t count)
-    {
+                 std::size_t count) {
       char *src0 = src[0];
       std::intptr_t src0_stride = src_stride[0];
       for (std::size_t i = 0; i < count; ++i) {
@@ -43,15 +42,13 @@ namespace nd {
     typedef DstType dst_type;
     typedef Src0Type src0_type;
 
-    void single(char *dst, char *const *src)
-    {
+    void single(char *dst, char *const *src) {
       *reinterpret_cast<dst_type *>(dst) =
           static_cast<dst_type>(*reinterpret_cast<dst_type *>(dst) + *reinterpret_cast<src0_type *>(src[0]));
     }
 
     void strided(char *dst, std::intptr_t dst_stride, char *const *src, const std::intptr_t *src_stride,
-                 std::size_t count)
-    {
+                 std::size_t count) {
       char *src0 = src[0];
       std::intptr_t src0_stride = src_stride[0];
       for (std::size_t i = 0; i < count; ++i) {
@@ -69,15 +66,13 @@ namespace nd {
     typedef bool1 dst_type;
     typedef Src0Type src0_type;
 
-    void single(char *dst, char *const *src)
-    {
+    void single(char *dst, char *const *src) {
       *reinterpret_cast<dst_type *>(dst) =
           *reinterpret_cast<dst_type *>(dst) + *reinterpret_cast<src0_type *>(src[0]) ? true : false;
     }
 
     void strided(char *dst, std::intptr_t dst_stride, char *const *src, const std::intptr_t *src_stride,
-                 std::size_t count)
-    {
+                 std::size_t count) {
       char *src0 = src[0];
       std::intptr_t src0_stride = src_stride[0];
       for (std::size_t i = 0; i < count; ++i) {
@@ -92,8 +87,7 @@ namespace nd {
   template <type_id_t DstTypeID, type_id_t Src0TypeID>
   struct compound_add_kernel_t
       : compound_add_kernel<typename type_of<DstTypeID>::type, typename type_of<Src0TypeID>::type,
-                            !is_lossless_assignable<DstTypeID, Src0TypeID>::value> {
-  };
+                            !is_lossless_assignable<DstTypeID, Src0TypeID>::value> {};
 
 } // namespace dynd::nd
 } // namespace dynd
